refactor : PR 코드리뷰 반영
