+++ src/app/interview/job-posting/[uuid]/complete/page.tsx
@@ -1,6 +1,6 @@
 import Header from '@/components/common/header/Header'
 import CompleteSection from '@/components/interview/CompleteSection'
-import { TOTAL_QUESTIONS } from '@/constants/interview'
+import { TOTAL_QUESTIONS } from '@/utils/interview'
 
 export default async function CompletePage({
   searchParams,
@@ -13,7 +13,10 @@ export default async function CompletePage({
   return (
     <main className="h-screen flex flex-col overflow-hidden">
       <Header />
-      <CompleteSection currentQuestion={currentQuestion} totalQuestions={TOTAL_QUESTIONS} />
+      <CompleteSection
+        currentQuestion={currentQuestion}
+        totalQuestions={TOTAL_QUESTIONS}
+      />
     </main>
   )
 }
