--- src/constants/interview.ts
@@ -1,1 +1,0 @@
-export const TOTAL_QUESTIONS = 5
